refactor(AppData): Move user data operations to AppData component.
Also move email verif screen logic to index.js
