+++ pages/index.js
@@ -4,6 +4,7 @@ import { useAuth } from 'use-auth0-hooks'
 import ApiUserSetup from '../components/ApiUserSetup'
 import ApiUserWelcome from '../components/ApiUserWelcome'
 import LogSummary from '../components/LogSummary'
+import VerifyEmailScreen from '../components/verify-email-screen'
 import UserList from '../components/UserList'
 import { AUTH0_SCOPE } from '../util/constants'
 
@@ -12,22 +13,27 @@ export default function Index (props) {
     adminUser,
     apiUser,
     createUser,
-    isUserFetched,
-    isUserRequested
+    isUserFetched
   } = props
   const { query } = useRouter()
-  const { isAuthenticated } = useAuth({
+  const { isAuthenticated, user } = useAuth({
     audience: process.env.AUTH0_AUDIENCE,
     scope: AUTH0_SCOPE
   })
-  // FIXME: isLoading appears to be broken in useAuth.
-  if (!isAuthenticated && !isUserRequested) {
+
+  if (!isAuthenticated) {
     return (
       <div>
         Please log in to view the Admin Dashboard.
       </div>
     )
   }
+
+  if (user && !user.email_verified) {
+    return <VerifyEmailScreen />
+  }
+
+  // FIXME: isLoading appears to be broken in useAuth.
   if (!isUserFetched) {
     return <div>Loading...</div>
   }
